revisi rm

+++ rm.c
@@ -1,17 +1,23 @@
-#include<stdio.h>
-#include <dirent.h>
-#include <sys/stat.h>
-#include<malloc.h>
-#include<string.h>
-#include <sys/types.h>
-#include<errno.h>
-#include <fcntl.h>
-#include <unistd.h>
+#include "types.h"
+#include "stat.h"
+#include "user.h"
 
 int main(int argc, char *argv[])
 {
+
+
+
+	if(argc<2)
+	{
+		printf(2,"Need 2 Arguments\n");
+		exit();
+	}
 	
-	unlink(argv[1]);
+	else
+	{
+		unlink(argv[1]);
+		exit();
+	}
+
 
 } 
-
